Gestion Erreur côté Panier ✔️

+++ Includes/removeItemCheckout.php
@@ -1,27 +1,22 @@
 <?php
-    session_start();
+    require_once 'SessionChecker.php';
     require_once 'dbh.php';
 
-    if (isset($_SESSION["Id"])){
-        if(isset($_GET['item'])){
-        $params = array($_SESSION['Id'], $_GET['item']);
+    if (isset($_SESSION["Id"]) && isset($_GET['item'])){
+    $params = array($_SESSION['Id'], $_GET['item']);
 
-        $sql = "delete from Panier where idJoueur = ? and idItem = ?";
+    $sql = "delete from Panier where idJoueur = ? and idItem = ?";
 
-        $stmt = sqlsrv_query($conn, $sql, $params);
-        
-        sqlsrv_close($conn);
+    $stmt = sqlsrv_query($conn, $sql, $params);
+    
+    sqlsrv_close($conn);
 
+    header('Location:../Panier.php');
+    exit();
+    }
+    else {
         header('Location:../Panier.php');
-        exit();
-        }
-        else {
-            header('Location:../Panier.php');
-            $_SESSION['removeCheckoutError'] = 'Impossible d\'enlever l\'item';
-            exit();
-        }
-    }else{
-        header('Location:../loginForm.php');
+        $_SESSION['removeCheckoutError'] = 'Impossible d\'enlever l\'item';
         exit();
     }
 ?>
